Adjusting package structure to move resources into api/resources
