+++ src/main/java/marquez/MarquezApp.java
@@ -13,6 +13,8 @@ import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import marquez.api.resources.DatasetResource;
 import marquez.api.resources.HealthResource;
+import marquez.api.resources.JobResource;
+import marquez.api.resources.NamespaceResource;
 import marquez.api.resources.PingResource;
 import marquez.core.mappers.ResourceExceptionMapper;
 import marquez.core.services.JobService;
@@ -23,8 +25,6 @@ import marquez.dao.JobVersionDAO;
 import marquez.dao.NamespaceDAO;
 import marquez.dao.RunArgsDAO;
 import marquez.db.DatasetDao;
-import marquez.resources.JobResource;
-import marquez.resources.NamespaceResource;
 import marquez.service.DatasetService;
 import org.flywaydb.core.Flyway;
 import org.flywaydb.core.api.FlywayException;
