+++ src/test/java/marquez/api/JobResourceTest.java
@@ -20,12 +20,12 @@ import java.util.UUID;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import marquez.api.resources.JobResource;
 import marquez.core.exceptions.UnexpectedException;
 import marquez.core.mappers.ResourceExceptionMapper;
 import marquez.core.models.Generator;
 import marquez.core.services.JobService;
 import marquez.core.services.NamespaceService;
-import marquez.resources.JobResource;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
